+++ src/main/java/si/matjazcerkvenik/alertmonitor/webhook/TargetBean.java
@@ -7,12 +7,13 @@ import si.matjazcerkvenik.alertmonitor.model.Target;
 import javax.annotation.PostConstruct;
 import javax.faces.bean.ManagedBean;
 import javax.faces.bean.RequestScoped;
+import javax.faces.bean.ViewScoped;
 import javax.faces.context.FacesContext;
 import java.util.*;
 import java.util.stream.Collectors;
 
 @ManagedBean
-@RequestScoped
+@ViewScoped
 public class TargetBean {
 
     private Target target;
@@ -36,6 +37,20 @@ public class TargetBean {
     public List<DNotification> getTargetActiveAlarms() {
         List<DNotification> list = new ArrayList<DNotification>(DAO.getInstance().getActiveAlerts().values());
         List<DNotification> result = list.stream()
+                .filter(notif -> checkAlert(notif))
+                .collect(Collectors.toList());
+        Collections.sort(result, new Comparator<DNotification>() {
+            @Override
+            public int compare(DNotification lhs, DNotification rhs) {
+                // -1 - less than, 1 - greater than, 0 - equal, all inversed for descending
+                return lhs.getTimestamp() > rhs.getTimestamp() ? -1 : (lhs.getTimestamp() < rhs.getTimestamp()) ? 1 : 0;
+            }
+        });
+        return result;
+    }
+
+    public List<DNotification> getTargetJournalAlarms() {
+        List<DNotification> result = DAO.getInstance().getJournal().stream()
                 .filter(notif -> checkAlert(notif))
                 .collect(Collectors.toList());
         Collections.sort(result, new Comparator<DNotification>() {