+++ src/main/java/si/matjazcerkvenik/alertmonitor/model/alertmanager/AlertmanagerResyncMetricObject.java
@@ -1,0 +1,22 @@
+package si.matjazcerkvenik.alertmonitor.model.alertmanager;
+
+public class AlertmanagerResyncMetricObject {
+
+    private AlertmanagerResyncAlert metric;
+
+    public AlertmanagerResyncAlert getMetric() {
+        return metric;
+    }
+
+    public void setMetric(AlertmanagerResyncAlert metric) {
+        this.metric = metric;
+    }
+
+    @Override
+    public String toString() {
+        return "AlertmanagerResyncMetricObject{" +
+                "metric=" + metric +
+                '}';
+    }
+}
+