+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.5.1-SNAPSHOT
 
 * [CHANGE] Target view shows also targets with no active alerts (targets with at least one record in journal).
+* [ENHANCEMENT] Instance view shows active alerts and journal
 * [ENHANCEMENT] Added `clearTimestamp` and `clearUid` to alerts; timestamp when alert was cleared
 * [ENHANCEMENT] Added some javadoc